Refactor asset parts, fix single image display on click.
