+++ app/src/main/res/layout-sw600dp-port/message_audio_asset.xml
@@ -17,9 +17,9 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 -->
-<com.waz.zclient.messages.parts.AudioAssetPartView xmlns:android="http://schemas.android.com/apk/res/android"
-                                                   android:layout_width="@dimen/content__width"
-                                                   android:layout_height="@dimen/content__audio_message__view_height"
-                                                   android:layout_gravity="center"
-                                                   android:paddingEnd="@dimen/content__row__padding_right"
+<com.waz.zclient.messages.parts.assets.AudioAssetPartView xmlns:android="http://schemas.android.com/apk/res/android"
+                                                          android:layout_width="@dimen/content__width"
+                                                          android:layout_height="@dimen/content__audio_message__view_height"
+                                                          android:layout_gravity="center"
+                                                          android:paddingEnd="@dimen/content__row__padding_right"
     />
