+++ app/src/main/res/layout/message_audio_asset.xml
@@ -18,9 +18,9 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 -->
-<com.waz.zclient.messages.parts.AudioAssetPartView xmlns:android="http://schemas.android.com/apk/res/android"
-                                                   android:layout_width="match_parent"
-                                                   android:layout_height="@dimen/content__audio_message__view_height"
-                                                   android:layout_marginEnd="@dimen/content__padding_right"
-                                                   android:layout_marginStart="@dimen/content__padding_left"
+<com.waz.zclient.messages.parts.assets.AudioAssetPartView xmlns:android="http://schemas.android.com/apk/res/android"
+                                                          android:layout_width="match_parent"
+                                                          android:layout_height="@dimen/content__audio_message__view_height"
+                                                          android:layout_marginEnd="@dimen/content__padding_right"
+                                                          android:layout_marginStart="@dimen/content__padding_left"
     />
