+++ app/src/main/res/layout/message_file_asset.xml
@@ -17,10 +17,10 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 -->
-<com.waz.zclient.messages.parts.FileAssetPartView xmlns:android="http://schemas.android.com/apk/res/android"
-                                                  android:layout_width="match_parent"
-                                                  android:layout_height="@dimen/content__audio_message__view_height"
-                                                  android:layout_marginEnd="@dimen/content__padding_right"
-                                                  android:layout_marginStart="@dimen/content__padding_left"
-                                                  android:paddingEnd="@dimen/wire__padding__12"
+<com.waz.zclient.messages.parts.assets.FileAssetPartView xmlns:android="http://schemas.android.com/apk/res/android"
+                                                         android:layout_width="match_parent"
+                                                         android:layout_height="@dimen/content__audio_message__view_height"
+                                                         android:layout_marginEnd="@dimen/content__padding_right"
+                                                         android:layout_marginStart="@dimen/content__padding_left"
+                                                         android:paddingEnd="@dimen/wire__padding__12"
     />
