+++ app/src/main/res/layout/message_file_asset_content.xml
@@ -32,7 +32,7 @@
         android:layout_height="match_parent"
         >
 
-        <com.waz.zclient.views.AssetActionButtonNew
+        <com.waz.zclient.messages.parts.assets.AssetActionButton
             android:id="@+id/action_button"
             android:layout_width="@dimen/content__audio_message__button__size"
             android:layout_height="@dimen/content__audio_message__button__size"
