+++ app/src/main/scala/com/waz/zclient/messages/parts/assets/AssetPart.scala
@@ -1,0 +1,126 @@
+package com.waz.zclient.messages.parts.assets
+
+import android.view.View
+import android.widget.TextView
+import com.waz.model.{AssetData, Dim2, MessageContent, MessageData}
+import com.waz.threading.Threading
+import com.waz.utils.events.Signal
+import com.waz.utils.returning
+import com.waz.zclient.controllers.AssetsController
+import com.waz.zclient.messages.MessageView.MsgBindOptions
+import com.waz.zclient.messages.MessageViewPart
+import com.waz.zclient.messages.parts.assets.DeliveryState.OtherUploading
+import com.waz.zclient.utils.ContextUtils._
+import com.waz.zclient.utils.{StringUtils, _}
+import com.waz.zclient.views.ImageAssetDrawable
+import com.waz.zclient.views.ImageAssetDrawable.Padding
+import com.waz.zclient.views.ImageController.WireImage
+import com.waz.zclient.{R, ViewHelper}
+import org.threeten.bp.Duration
+
+trait AssetPart extends View with MessageViewPart with ViewHelper {
+  val controller = inject[AssetsController]
+
+  val message = Signal[MessageData]()
+  val asset = controller.assetSignal(message)
+  val deliveryState = DeliveryState(message, asset)
+  val completed = deliveryState.map(_ == DeliveryState.Complete)
+
+  val progressDots = new AssetBackground(deliveryState.map(_ == OtherUploading))
+  setBackground(progressDots)
+
+  override def set(msg: MessageData, part: Option[MessageContent], opts: MsgBindOptions): Unit = {
+    message ! msg
+  }
+}
+
+
+trait ContentAssetPart extends AssetPart {
+  def inflate(): Unit
+
+  inflate()
+  private val content: View = findById(R.id.content)
+
+  //toggle content visibility to show only progress dot background if other side is uploading asset
+  deliveryState.map {
+    case OtherUploading => false
+    case _ => true
+  }.on(Threading.Ui)(content.setVisible)
+}
+
+
+trait ActionableAssetPart extends ContentAssetPart {
+  protected val assetActionButton: AssetActionButton = findById(R.id.action_button)
+
+  override def set(msg: MessageData, part: Option[MessageContent], opts: MsgBindOptions): Unit = {
+    super.set(msg, part, opts)
+    assetActionButton.message.publish(msg, Threading.Ui)
+  }
+}
+
+trait PlayableAsset extends ActionableAssetPart {
+  val duration = asset.map(_._1).map {
+    case AssetData.WithDuration(d) => d
+    case _ => Duration.ZERO
+  }
+  val formattedDuration = duration.map(d => StringUtils.formatTimeSeconds(d.getSeconds))
+
+  protected val durationView: TextView = findById(R.id.duration)
+
+  //TODO there is more logic for what text to display in video views, but it doesn't seem to be used - confirm
+  formattedDuration.on(Threading.Ui)(durationView.setText)
+
+}
+
+
+trait ImageLayoutAssetPart extends AssetPart {
+  protected val imageDim = message map { _.imageDimensions.getOrElse(Dim2(1, 1)) }
+  protected val viewWidth = Signal[Int]()
+
+  private lazy val contentPadding = getDimenPx(R.dimen.content__padding_left)
+
+  val imageDrawable = new ImageAssetDrawable(message map { m => WireImage(m.assetId) })
+
+  val displaySize = for {
+    w <- viewWidth
+    Dim2(imW, imH) <- imageDim
+  } yield {
+    val pxW = toPx(imW)
+    val centered = w - 2 * contentPadding
+    val padded = w - contentPadding
+    val width =
+      if (imH > imW) math.min(pxW, centered)
+      else if (pxW >= padded) w
+      else if (pxW >= centered) centered
+      else pxW
+
+    Dim2(width, imH * width / imW)
+  }
+
+  val padding = for {
+    w <- viewWidth
+    Dim2(dW, dH) <- displaySize
+  } yield {
+    if (dW >= w) Padding.Empty
+    else {
+      val left = if (getLayoutDirection == View.LAYOUT_DIRECTION_LTR) contentPadding else w - contentPadding - dW
+      Padding(left, 0, w - dW - left, 0)
+    }
+  }
+
+  padding { imageDrawable.padding ! _ }
+
+  displaySize.map(_.height) { h =>
+    setLayoutParams(returning(getLayoutParams)(_.height = h))
+  }
+
+  override def set(msg: MessageData, part: Option[MessageContent], opts: MsgBindOptions): Unit = {
+    super.set(msg, part, opts)
+    viewWidth.mutateOrDefault(identity, opts.widthHint)
+  }
+
+  override def onLayout(changed: Boolean, left: Int, top: Int, right: Int, bottom: Int): Unit = {
+    super.onLayout(changed, left, top, right, bottom)
+    viewWidth ! (right - left)
+  }
+}
